GUI refactor in anticipation of inventory screen, fixed axe bug
