+++ src/gui/description.py
@@ -26,13 +26,10 @@ class Description:
         # description
         self.description_rect = pygame.Rect(0, 0, 600, 400)
         self.description_rect.topright = self.rect.topright
-        self.description_surface = pygame.Surface(self.description_rect.size)
-        self.description_surface.set_colorkey('green')
-
+        self.description_surface = pygame.Surface(self.description_rect.size, pygame.SRCALPHA)
         # slider
-        self.description_slider_surface = pygame.Surface((600, 600))
+        self.description_slider_surface = pygame.Surface((600, 600), pygame.SRCALPHA)
         self.description_slider_rect = self.description_surface.get_rect()
-        self.description_slider_surface.set_colorkey('green')
 
     # events
     def handle_events(self, event):
@@ -55,8 +52,7 @@ class Description:
 
     # draw
     def make_surface_transparent(self):
-        self.description_surface.fill('green')
-        self.description_slider_surface.fill('green')
+        self.description_surface.fill(pygame.Color(0, 0, 0, 0))
 
     def draw_slider_bar(self):
         height1 = self.description_slider_surface.get_height()
