+++ src/gui/general_menu.py
@@ -7,39 +7,21 @@ from src.gui.components import Button
 from src.enums import GameState
 from pygame.mouse import get_pressed as mouse_buttons
 from pygame.math import Vector2 as vector
+from src.gui.abstract_menu import AbstractMenu
 
 
 _SCREEN_CENTER = (SCREEN_WIDTH // 2, SCREEN_HEIGHT // 2)
 
 
-class GeneralMenu:
+class GeneralMenu(AbstractMenu):
     def __init__(self,  title, options, switch, size, center=vector()):
-        # general setup
-        self.display_surface = pygame.display.get_surface()
-        self.buttons_surface = pygame.Surface(size)
-        self.buttons_surface.set_colorkey('green')
-        self.font = pygame.font.Font(resource_path('font/LycheeSoda.ttf'), 30)
-        self.title = title
+        super().__init__(title, size, center)
 
-        # rect
-        self.center = center
-        self.size = size
-        self.rect = None
-        self.rect_setup()
-
-        # buttons
-        self.pressed_button = None
         self.options = options
-        self.buttons = []
         self.button_setup()
 
         # switch
         self.switch_screen = switch
-
-    # setup
-    def rect_setup(self):
-        self.rect = pygame.Rect((0, 0), self.size)
-        self.rect.center = self.center or _SCREEN_CENTER
 
     def button_setup(self):
         # button setup
@@ -61,87 +43,9 @@ class GeneralMenu:
             self.buttons.append(button)
             generic_button_rect = rect.move(0, button_height + space)
 
-    # events
-    def event_loop(self):
-        self.mouse_hover()
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                self.quit_game()
-
-            self.click(event)
-            self.handle_events(event)
-
-    def handle_events(self, event):
-        pass
-
-    def get_hovered_button(self):
-        for button in self.buttons:
-            if button.mouse_hover():
-                return button
-        return None
-
-    def click(self, event):
-        if event.type == pygame.MOUSEBUTTONDOWN and mouse_buttons()[0]:
-            self.pressed_button = self.get_hovered_button()
-            if self.pressed_button:
-                self.pressed_button.start_press_animation()
-
-        if event.type == pygame.MOUSEBUTTONUP:
-            if self.pressed_button:
-                self.pressed_button.start_release_animation()
-
-                if self.pressed_button.mouse_hover():
-                    self.button_action(self.pressed_button.text)
-                    pygame.mouse.set_cursor(pygame.SYSTEM_CURSOR_ARROW)
-
-                self.pressed_button = None
-
-    def mouse_hover(self):
-        for button in self.buttons:
-            if button.hover_active:
-                pygame.mouse.set_cursor(pygame.SYSTEM_CURSOR_HAND)
-                return
-        pygame.mouse.set_cursor(pygame.SYSTEM_CURSOR_ARROW)
-
     def button_action(self, text):
         if text == 'Play':
+            pygame.mouse.set_cursor(pygame.SYSTEM_CURSOR_ARROW)
             self.switch_screen(GameState.LEVEL)
         if text == 'Quit':
             self.quit_game()
-
-    def quit_game(self):
-        pygame.quit()
-        sys.exit()
-
-    # draw
-    def draw_title(self):
-        text_surf = self.font.render(self.title, False, 'Black')
-        midtop = (SCREEN_WIDTH / 2, SCREEN_HEIGHT / 20)
-        text_rect = text_surf.get_frect(midtop=midtop)
-
-        bg_rect = pygame.Rect((0, 0), (200, 50))
-        bg_rect.center = text_rect.center
-
-        pygame.draw.rect(self.display_surface, 'White', bg_rect, 0, 4)
-        self.display_surface.blit(text_surf, text_rect)
-
-    def draw_buttons(self):
-        self.buttons_surface.fill('green')
-        for button in self.buttons:
-            button.draw(self.display_surface)
-        self.display_surface.blit(self.buttons_surface, self.rect.topleft)
-
-    def draw(self):
-        self.display_surface.fill('cadetblue')
-        self.draw_title()
-        self.draw_buttons()
-
-    # update
-    def update_buttons(self, dt):
-        for button in self.buttons:
-            button.update(dt)
-
-    def update(self, dt):
-        self.event_loop()
-        self.update_buttons(dt)
-        self.draw()
